refactor(commands): extract group tests to sibling file
